Update seeder for supported ruby/rails versions
This commit updates the CI configuration for this gem to test against
all currently supported versions of Ruby and ActiveRecord.

It also switches CI from travis-ci.org to GitHub Actions.
