--- .travis.yml
@@ -1,45 +1,0 @@
-language: ruby
-cache: bundler
-services:
-  - mysql
-rvm:
-  - 2.1.10
-  - 2.2.7
-  - 2.3.4
-  - 2.4.1
-  - 2.5.8
-  - 2.6.6
-  - 2.7.1
-gemfile:
-  - gemfiles/5.0.gemfile
-  - gemfiles/5.1.gemfile
-  - gemfiles/5.2.gemfile
-  - gemfiles/6.0.gemfile
-matrix:
-  exclude:
-    - rvm: 2.2.7
-      gemfile: gemfiles/6.0.gemfile
-
-    - rvm: 2.1.10
-      gemfile: gemfiles/5.0.gemfile
-    - rvm: 2.1.10
-      gemfile: gemfiles/5.1.gemfile
-    - rvm: 2.1.10
-      gemfile: gemfiles/5.2.gemfile
-    - rvm: 2.1.10
-      gemfile: gemfiles/6.0.gemfile
-
-    - rvm: 2.3.4
-      gemfile: gemfiles/6.0.gemfile
-
-    - rvm: 2.4.1
-      gemfile: gemfiles/6.0.gemfile
-
-    - rvm: 2.7.1
-      gemfile: gemfiles/5.0.gemfile
-    - rvm: 2.7.1
-      gemfile: gemfiles/5.1.gemfile
-    - rvm: 2.7.1
-      gemfile: gemfiles/5.2.gemfile
-before_script:
-  - mysql -e 'create database seeder_test;'
