+++ Appraisals
@@ -1,13 +1,3 @@
-appraise '5.0' do
-  gem 'activerecord', '~> 5.0.0'
-  gem 'mysql2', '~> 0.4.0'
-end
-
-appraise '5.1' do
-  gem 'activerecord', '~>5.1.0'
-  gem 'mysql2', '~> 0.4.0'
-end
-
 appraise '5.2' do
   gem 'activerecord', '~> 5.2.0'
   gem 'mysql2', '~> 0.4.0'
@@ -16,4 +6,14 @@ end
 appraise '6.0' do
   gem 'activerecord', '~> 6.0.3.3'
   gem 'mysql2', '~> 0.4.0'
+end
+
+appraise '6.1' do
+  gem 'activerecord', '~> 6.1.0'
+  gem 'mysql2', '~> 0.5.3'
+end
+
+appraise '7.0' do
+  gem 'activerecord', '~> 7.0.0'
+  gem 'mysql2', '~> 0.5.3'
 end
