+++ gemfiles/6.1.gemfile
@@ -3,7 +3,7 @@
 source "https://rubygems.org"
 
 gem "appraisal"
-gem "activerecord", "~>5.1.0"
-gem "mysql2", "~> 0.4.0"
+gem "activerecord", "~> 6.1.0"
+gem "mysql2", "~> 0.5.3"
 
 gemspec path: "../"
