+++ gemfiles/7.0.gemfile
@@ -3,7 +3,7 @@
 source "https://rubygems.org"
 
 gem "appraisal"
-gem "activerecord", "~> 5.0.0"
-gem "mysql2", "~> 0.4.0"
+gem "activerecord", "~> 7.0.0"
+gem "mysql2", "~> 0.5.3"
 
 gemspec path: "../"
